Moved cat. class to cat. folder
