+++ frontend/src/app/categories/category-detail/category-detail.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, Input } from '@angular/core';
 import { ActivatedRoute } from "@angular/router";
 import { Location } from "@angular/common";
 
-import { Category } from "../../models"
+import { Category } from "../category"
 import { GenericService } from "../../generic.service";
 
 @Component({
